Bugfix: Resolve conflict issues

+++ src/app/(routes)/dashboard/page.tsx
@@ -4,15 +4,12 @@ import React, { useState, useEffect } from "react";
 import { useRouter } from "next/navigation";
 import Image from "next/image";
 import { students } from "@/utils/data";
+import Swal from "sweetalert2";
 
 function DashboardPage() {
-
-
   type Attendance = {
     [key: number]: string;
   };
-
-
 
   const [attendance, setAttendance] = useState<Attendance>({});
   const router = useRouter();
@@ -30,12 +27,29 @@ function DashboardPage() {
   }, [router]);
 
   const registerAttendance = (studentId: number) => {
-    const today = new Date().toISOString().split("T")[0];
-    const newAttendance = { ...attendance, [studentId]: today };
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You are about to register this student.",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, register it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        const today = new Date().toISOString().split("T")[0];
+        const newAttendance = { ...attendance, [studentId]: today };
 
-    setAttendance(newAttendance);
-    localStorage.setItem("attendance", JSON.stringify(newAttendance));
-    alert("Student has been registered");
+        setAttendance(newAttendance);
+        localStorage.setItem("attendance", JSON.stringify(newAttendance));
+
+        Swal.fire({
+          icon: "success",
+          title: "Registered",
+          text: "Student has been registered",
+        });
+      }
+    });
   };
 
   return (
@@ -61,7 +75,9 @@ function DashboardPage() {
                   <Image
                     src={student.picture}
                     alt={student.fullname}
-                    className="w-12 h-12 rounded-full"
+                    width={48} 
+                    height={48} 
+                    className="rounded-full"
                   />
                 </td>
                 <td className="py-2 px-4 border-b">{student.fullname}</td>
@@ -69,21 +85,21 @@ function DashboardPage() {
                 <td className="py-2 px-4 border-b">{student.course}</td>
                 <td className="py-2 px-4 border-b">{student.dept}</td>
                 <td className="py-2 px-4 border-b">
-                  {
-                    attendance[student.id]? <button
-                    onClick={() => alert("Already registered")}
-                    className="bg-yellow-500 text-white px-4 py-1 rounded hover:bg-yellow-700 transition duration-200"
-                  >
-                    Registered
-                  </button>:  <button
-                    onClick={() => registerAttendance(student.id)}
-                    className="bg-blue-500 text-white px-4 py-1 rounded hover:bg-blue-700 transition duration-200"
-                  >
-                    Mark Register
-                  </button>
-                  }
-                 
-                
+                  {attendance[student.id] ? (
+                    <button
+                      onClick={() => alert("Already registered")}
+                      className="bg-yellow-500 text-white px-4 py-1 rounded hover:bg-yellow-700 transition duration-200"
+                    >
+                      Registered
+                    </button>
+                  ) : (
+                    <button
+                      onClick={() => registerAttendance(student.id)}
+                      className="bg-blue-500 text-white px-4 py-1 rounded hover:bg-blue-700 transition duration-200"
+                    >
+                      Mark Register
+                    </button>
+                  )}
                 </td>
                 <td className="py-2 px-4 border-b">
                   {attendance[student.id] ? (
